Update to Spring Kafka 2.5.1 snapshot
The changes introduced in 0ac83e510f did not fix the
binary compatibility issue with Spring Kakfa 2.3.7
since the binder was compiling against a new version (2.5.0).
Spirng Kafka added the missing constructor and that fixed the
binary compatiblity issue with 2.3.7.

See this commit in Spring Kafka for reference:
https://github.com/spring-projects/spring-kafka/commit/b404785d8dca5cdb3a3b8b1c5b746141d66cf533

Updating to the latest snaphot (2.5.1).