Remove ThreadLocal in KafkaStreams health check
Remove an optimization that was added in Kafka Streams binder
health check where it caches DOWN status if one of the binder
already detects that the broker is down (since there are three
different binders in Kafka Streams). The caching was done using
a ThreadLocal which was causing issues when using Reactive libraries
in the application. Removing this optimization.

Applications can still reduce the time it takes for health check
in the event when the Kafka broker is down by using the health timeout
offered by the binder properties.

Resolves https://github.com/spring-cloud/spring-cloud-stream-binder-kafka/issues/939