GH-1107: Kafka metrics lag check
When KafkaBinderMetrics checks lags for offsets in partitions, if the partitions
have nothing committed yet, it unnecessarily tries to find the last committed offset.
This is causing some unncessary logging. Avoid this by only checking the committed
offsets in a partition, if the current end offset is greater than zero.

Resolves https://github.com/spring-cloud/spring-cloud-stream-binder-kafka/issues/1107